+++ pdf/lib/src/ttf_parser.dart
@@ -16,6 +16,17 @@
 
 part of pdf;
 
+class TtfGlyphInfo {
+  TtfGlyphInfo(this.index, this.data, this.compounds);
+
+  final int index;
+  final Uint8List data;
+  final List<int> compounds;
+
+  @override
+  String toString() => 'Glyph $index $compounds';
+}
+
 class TtfParser {
   TtfParser(this.bytes) {
     final int numTables = bytes.getUint16(4);
@@ -23,56 +34,61 @@ class TtfParser {
     for (int i = 0; i < numTables; i++) {
       final String name = utf8.decode(bytes.buffer.asUint8List(i * 16 + 12, 4));
       final int offset = bytes.getUint32(i * 16 + 20);
-      _tableOffsets[name] = offset;
+      final int size = bytes.getUint32(i * 16 + 24);
+      tableOffsets[name] = offset;
+      tableSize[name] = size;
     }
 
     _parseFontName();
     _parseCMap();
     _parseIndexes();
-    _parseGlyf();
+    _parseGlyphs();
   }
 
-  static const String _HEAD = 'head';
-  static const String _NAME = 'name';
-  static const String _HMTX = 'hmtx';
-  static const String _HHEA = 'hhea';
-  static const String _CMAP = 'cmap';
-  static const String _MAXP = 'maxp';
-  static const String _LOCA = 'loca';
-  static const String _GLYF = 'glyf';
+  static const String head_table = 'head';
+  static const String name_table = 'name';
+  static const String hmtx_table = 'hmtx';
+  static const String hhea_table = 'hhea';
+  static const String cmap_table = 'cmap';
+  static const String maxp_table = 'maxp';
+  static const String loca_table = 'loca';
+  static const String glyf_table = 'glyf';
 
   final ByteData bytes;
-  final Map<String, int> _tableOffsets = <String, int>{};
+  final Map<String, int> tableOffsets = <String, int>{};
+  final Map<String, int> tableSize = <String, int>{};
   String _fontName;
   final Map<int, int> charToGlyphIndexMap = <int, int>{};
   final List<int> glyphOffsets = <int>[];
   final Map<int, PdfFontMetrics> glyphInfoMap = <int, PdfFontMetrics>{};
 
-  int get unitsPerEm => bytes.getUint16(_tableOffsets[_HEAD] + 18);
+  int get unitsPerEm => bytes.getUint16(tableOffsets[head_table] + 18);
 
-  int get xMin => bytes.getInt16(_tableOffsets[_HEAD] + 36);
+  int get xMin => bytes.getInt16(tableOffsets[head_table] + 36);
 
-  int get yMin => bytes.getInt16(_tableOffsets[_HEAD] + 38);
+  int get yMin => bytes.getInt16(tableOffsets[head_table] + 38);
 
-  int get xMax => bytes.getInt16(_tableOffsets[_HEAD] + 40);
+  int get xMax => bytes.getInt16(tableOffsets[head_table] + 40);
 
-  int get yMax => bytes.getInt16(_tableOffsets[_HEAD] + 42);
+  int get yMax => bytes.getInt16(tableOffsets[head_table] + 42);
 
-  int get indexToLocFormat => bytes.getInt16(_tableOffsets[_HEAD] + 50);
+  int get indexToLocFormat => bytes.getInt16(tableOffsets[head_table] + 50);
 
-  int get ascent => bytes.getInt16(_tableOffsets[_HHEA] + 4);
+  int get ascent => bytes.getInt16(tableOffsets[hhea_table] + 4);
 
-  int get descent => bytes.getInt16(_tableOffsets[_HHEA] + 6);
+  int get descent => bytes.getInt16(tableOffsets[hhea_table] + 6);
 
-  int get numOfLongHorMetrics => bytes.getInt16(_tableOffsets[_HHEA] + 34);
+  int get numOfLongHorMetrics => bytes.getInt16(tableOffsets[hhea_table] + 34);
 
-  int get numGlyphs => bytes.getInt16(_tableOffsets[_MAXP] + 4);
+  int get numGlyphs => bytes.getInt16(tableOffsets[maxp_table] + 4);
 
   String get fontName => _fontName;
 
+  bool get unicode => bytes.getUint32(0) == 0x10000;
+
   // https://developer.apple.com/fonts/TrueType-Reference-Manual/RM06/Chap6name.html
   void _parseFontName() {
-    final int basePosition = _tableOffsets[_NAME];
+    final int basePosition = tableOffsets[name_table];
     final int count = bytes.getUint16(basePosition + 2);
     final int stringOffset = bytes.getUint16(basePosition + 4);
     int pos = basePosition + 6;
@@ -105,7 +121,7 @@ class TtfParser {
   }
 
   void _parseCMap() {
-    final int basePosition = _tableOffsets[_CMAP];
+    final int basePosition = tableOffsets[cmap_table];
     final int numSubTables = bytes.getUint16(basePosition + 2);
     for (int i = 0; i < numSubTables; i++) {
       final int offset = bytes.getUint32(basePosition + i * 8 + 8);
@@ -190,7 +206,7 @@ class TtfParser {
   }
 
   void _parseIndexes() {
-    final int basePosition = _tableOffsets[_LOCA];
+    final int basePosition = tableOffsets[loca_table];
     final int numGlyphs = this.numGlyphs;
     if (indexToLocFormat == 0) {
       for (int i = 0; i < numGlyphs; i++) {
@@ -204,9 +220,9 @@ class TtfParser {
   }
 
   /// https://developer.apple.com/fonts/TrueType-Reference-Manual/RM06/Chap6glyf.html
-  void _parseGlyf() {
-    final int baseOffset = _tableOffsets[_GLYF];
-    final int hmtxOffset = _tableOffsets[_HMTX];
+  void _parseGlyphs() {
+    final int baseOffset = tableOffsets[glyf_table];
+    final int hmtxOffset = tableOffsets[hmtx_table];
     final int unitsPerEm = this.unitsPerEm;
     int glyphIndex = 0;
     for (int offset in glyphOffsets) {
@@ -227,5 +243,118 @@ class TtfParser {
           advanceWidth: advanceWidth);
       glyphIndex++;
     }
+  }
+
+  /// http://stevehanov.ca/blog/?id=143
+  TtfGlyphInfo readGlyph(int index) {
+    assert(index != null);
+    assert(index < glyphOffsets.length);
+
+    final int start = tableOffsets[glyf_table] + glyphOffsets[index];
+
+    final int numberOfContours = bytes.getInt16(start);
+    assert(numberOfContours >= -1);
+
+    if (numberOfContours == -1) {
+      return _readCompoundGlyph(index, start, start + 10);
+    } else {
+      return _readSimpleGlyph(index, start, start + 10, numberOfContours);
+    }
+  }
+
+  TtfGlyphInfo _readSimpleGlyph(
+      int glyph, int start, int offset, int numberOfContours) {
+    const int X_IS_BYTE = 2;
+    const int Y_IS_BYTE = 4;
+    const int REPEAT = 8;
+    const int X_DELTA = 16;
+    const int Y_DELTA = 32;
+
+    int numPoints = 1;
+
+    for (int i = 0; i < numberOfContours; i++) {
+      numPoints = math.max(numPoints, bytes.getUint16(offset) + 1);
+      offset += 2;
+    }
+
+    // skip over intructions
+    offset += bytes.getUint16(offset) + 2;
+
+    if (numberOfContours == 0) {
+      return TtfGlyphInfo(
+        glyph,
+        Uint8List.view(bytes.buffer, start, offset - start),
+        <int>[],
+      );
+    }
+
+    final List<int> flags = <int>[];
+
+    for (int i = 0; i < numPoints; i++) {
+      final int flag = bytes.getUint8(offset++);
+      flags.add(flag);
+
+      if (flag & REPEAT != 0) {
+        int repeatCount = bytes.getUint8(offset++);
+        assert(repeatCount > 0);
+        i += repeatCount;
+        while (repeatCount-- > 0) {
+          flags.add(flag);
+        }
+      }
+    }
+
+    int byteFlag = X_IS_BYTE;
+    int deltaFlag = X_DELTA;
+    for (int a = 0; a < 2; a++) {
+      for (int i = 0; i < numPoints; i++) {
+        final int flag = flags[i];
+        if (flag & byteFlag != 0) {
+          offset++;
+        } else if (~flag & deltaFlag != 0) {
+          offset += 2;
+        }
+      }
+      byteFlag = Y_IS_BYTE;
+      deltaFlag = Y_DELTA;
+    }
+
+    return TtfGlyphInfo(
+      glyph,
+      Uint8List.view(bytes.buffer, start, offset - start),
+      <int>[],
+    );
+  }
+
+  TtfGlyphInfo _readCompoundGlyph(int glyph, int start, int offset) {
+    const int ARG_1_AND_2_ARE_WORDS = 1;
+    const int MORE_COMPONENTS = 32;
+    const int WE_HAVE_INSTRUCTIONS = 256;
+
+    final List<int> components = <int>[];
+    bool hasInstructions = false;
+    int flags = MORE_COMPONENTS;
+
+    while (flags & MORE_COMPONENTS != 0) {
+      flags = bytes.getUint16(offset);
+      final int glyphIndex = bytes.getUint16(offset + 2);
+      offset += (flags & ARG_1_AND_2_ARE_WORDS != 0) ? 8 : 6;
+
+      components.add(glyphIndex);
+      if (flags & WE_HAVE_INSTRUCTIONS != 0) {
+        assert(!hasInstructions); // Not already set
+        hasInstructions = true;
+      }
+    }
+
+    if (hasInstructions) {
+      offset += bytes.getUint16(offset) + 2;
+    }
+
+    return TtfGlyphInfo(
+      glyph,
+      Uint8List.view(bytes.buffer, start, offset - start),
+      components,
+    );
   }
 }