+++ pdf/test/ttf_test.dart
@@ -65,10 +65,11 @@ void main() {
 
     final PdfGraphics g = page.getGraphics();
     int top = 0;
-    const String s = 'Hello ';
+    const String s = 'Hello Lukáča ';
 
     printTextTtf(g, s, File('open-sans.ttf'), 30.0 + 30.0 * top++);
     printTextTtf(g, s, File('roboto.ttf'), 30.0 + 30.0 * top++);
+    printTextTtf(g, s, File('noto-sans.ttf'), 30.0 + 30.0 * top++);
 
     final File file = File('ttf.pdf');
     file.writeAsBytesSync(pdf.save());