+++ pdf/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.3.5
 * Add some color functions
 * Remove color constants from PdfColor, use PdfColors
+* Add TTF Font SubSetting
+* Add Unicode support for TTF Fonts
 
 # 1.3.4
 * Add available dimensions for PdfPageFormat